# docs/openapi/user-service.yml
openapi: 3.0.3
info:
  title: User Service Contract
  version: 1.0.0
  description: Gateway 연동 기준 User Service 핵심 계약이다.
servers:
  - url: http://user-service:8082
...
paths:
  /users/signup:
    post:
      tags: [PublicUser]
      summary: 공개 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSignupRequest'
            example:
              email: user@example.com
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponseEnvelope'
              example:
                httpStatus: 201
                success: true
                message: 회원가입 성공
                code: 3001
                data:
                  user:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    email: user@example.com
                    role: USER
                    status: ACTIVE
                    createdAt: '2026-04-01T12:00:00'
                    updatedAt: '2026-04-01T12:00:00'
                    userSocialList: []
        default:
          $ref: '#/components/responses/UserError'
  /users/me:
    get:
      tags: [PublicUser]
      summary: 내 정보 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponseEnvelope'
              example:
                httpStatus: 200
                success: true
                message: 내 사용자 정보 조회 성공
                code: 3000
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  email: user@example.com
                  role: USER
                  status: ACTIVE
                  createdAt: '2026-04-01T12:00:00'
                  updatedAt: '2026-04-01T12:00:00'
                  userSocialList: []
        default:
          $ref: '#/components/responses/UserError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UserError:
      description: User service error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserErrorResponseEnvelope'
          example:
            httpStatus: 400
            success: false
            message: 잘못된 요청입니다.
            code: 7000
            data: null
  schemas:
    UserSignupRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...
    UserSocialResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        provider:
          $ref: '#/components/schemas/UserSocialType'
        providerUserId:
          type: string
        email:
          type: string
          format: email
        socialType:
          $ref: '#/components/schemas/UserSocialType'
        providerId:
          type: string
    UserDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userSocialList:
          type: array
          items:
            $ref: '#/components/schemas/UserSocialResponse'
    UserRole:
      type: string
      description: API에서는 enum 이름만 사용한다.
      enum:
        - SUPER_ADMIN
        - ADMIN
        - USER
        - GUEST
    UserStatus:
      type: string
      description: API에서는 enum 이름만 사용하며, DB에는 A/P/S/D 코드로 저장한다.
      enum:
        - ACTIVE
        - PENDING
        - SUSPENDED
        - DELETED
    UserSocialType:
      type: string
      description: API에서는 enum 이름만 사용하며, DB에는 DEF/GGL/KKO/GHB 코드로 저장한다.
      enum:
        - DEFAULT
        - GOOGLE
        - KAKAO
        - GITHUB
    UserCreateResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserDetailResponse'
    UserDetailResponseEnvelope:
      type: object
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        message:
          type: string
        code:
          type: integer
        data:
          $ref: '#/components/schemas/UserDetailResponse'
    UserCreateResponseEnvelope:
      type: object
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        message:
          type: string
        code:
          type: integer
        data:
          $ref: '#/components/schemas/UserCreateResponse'
...
    UserErrorResponseEnvelope:
      type: object
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        message:
          type: string
        code:
          type: integer
        data:
          nullable: true